Start cross-platform error handling in ipc
