+++ exp/pm/internal/ipc/internal/flisten/error.go
@@ -11,23 +11,9 @@ var (
 	ErrFileNotExist = errors.New("flisten file does not exist")
 )
 
-func asInUse(err error) error {
-	if errors.Is(err, syscall.EADDRINUSE) {
-		return ErrInUse
-	}
-	return err
-}
-
 func asFileNotExist(err error) error {
 	if errors.Is(err, syscall.ENOENT) {
 		return ErrFileNotExist
-	}
-	return err
-}
-
-func asConnRefused(err error) error {
-	if errors.Is(err, syscall.ECONNREFUSED) {
-		return ErrConnRefused
 	}
 	return err
 }
