+++ exp/pm/internal/ipc/internal/flisten/error_unix.go
@@ -1,0 +1,23 @@
+//go:build linux || darwin
+// +build linux darwin
+
+package flisten
+
+import (
+	"errors"
+	"syscall"
+)
+
+func asInUse(err error) error {
+	if errors.Is(err, syscall.EADDRINUSE) {
+		return ErrInUse
+	}
+	return err
+}
+
+func asConnRefused(err error) error {
+	if errors.Is(err, syscall.ECONNREFUSED) {
+		return ErrConnRefused
+	}
+	return err
+}
