+++ exp/pm/internal/ipc/internal/flisten/error_windows.go
@@ -1,0 +1,22 @@
+package flisten
+
+import (
+	"errors"
+
+	"golang.org/x/sys/windows"
+)
+
+func asInUse(err error) error {
+	if errors.Is(err, windows.WSAEADDRINUSE) {
+		return ErrInUse
+	}
+
+	return err
+}
+
+func asConnRefused(err error) error {
+	if errors.Is(err, windows.WSAECONNREFUSED) {
+		return ErrConnRefused
+	}
+	return err
+}
