+++ exp/pm/internal/svcctl/svcctl_unix.go
@@ -1,0 +1,8 @@
+//go:build linux || darwin
+// +build linux darwin
+
+package svcctl
+
+var (
+	fileExt = ""
+)
