+++ exp/pm/internal/svcctl/svcctl_windows.go
@@ -1,0 +1,5 @@
+package svcctl
+
+var (
+	fileExt = ".exe"
+)
